Separate Feeds Tests from Messages Tests
