+++ tests/integration/Social/Services/FeedsCest.php
@@ -4,16 +4,12 @@ namespace Kanvas\Packages\Tests\Integration\Social\Service;
 
 use IntegrationTester;
 use Kanvas\Packages\Social\Models\Channels;
-use Kanvas\Packages\Social\Models\Interactions as ModelsInteractions;
-use Kanvas\Packages\Social\Services\Feeds;
+use Kanvas\Packages\Social\Services\Messages as MessagesService;
 use Kanvas\Packages\Social\Services\MessageTypes;
-use Kanvas\Packages\Social\Services\Reactions;
 use Kanvas\Packages\Test\Support\Models\Users;
 use Kanvas\Packages\Social\Models\Messages;
 use Kanvas\Packages\Social\Services\Distributions;
-use Kanvas\Packages\Social\Services\Interactions;
 use Kanvas\Packages\Test\Support\Models\Lead;
-use Kanvas\Packages\Test\Support\Models\MessageObject;
 
 class FeedsCest
 {
@@ -41,7 +37,7 @@ class FeedsCest
             'text' => 'Test some messages'
         ];
         
-        $this->message = Feeds::create(new Users(), 'memo', $text);
+        $this->message = MessagesService::create(new Users(), 'memo', $text);
     }
 
     /**
@@ -65,55 +61,6 @@ class FeedsCest
         Distributions::sendToChannelFeed($this->channel, $this->message);
 
     }
-        
-    /**
-     * Test add comment
-     *
-     * @param UnitTester $I
-     * @before createMessageType
-     * @return void
-     */
-    public function createMessage(IntegrationTester $I): void
-    {
-        $text = [
-            'text' => 'This is test text for testing'
-        ];
-        
-        $feed = Feeds::create(new Users(), 'memo', $text, new MessageObject());
-
-        $I->assertNotNull($feed->getId());
-    }
-
-    /**
-     * Test Message Reactions
-     *
-     * @param IntegrationTester $I
-     * @before messageTestCreation
-     * @return void
-     */
-    public function messageReaction(IntegrationTester $I): void
-    {
-        $reaction = Reactions::addMessageReaction('confuse', new Users(), $this->message);
-
-        $I->assertEquals($this->message->getId(), $reaction->entity_id);
-
-        $reaction = Reactions::addMessageReaction('☺', new Users(), $this->message);
-        $I->assertEquals($this->message->getId(), $reaction->entity_id);
-    }
-
-    /**
-     * Test Message Interaction
-     *
-     * @param IntegrationTester $I
-     * @before messageTestCreation
-     * @return void
-     */
-    public function messageInteraction(IntegrationTester $I): void
-    {
-        $I->assertFalse(
-            Interactions::add(new Users(), $this->message, ModelsInteractions::REACT)
-        );
-    }
 
     /**
      * Test get feeds by channel
@@ -126,18 +73,5 @@ class FeedsCest
     public function getFeedsByChannel(IntegrationTester $I): void
     {
         
-    }
-
-    /**
-     * Test message and feed delete
-     *
-     * @param IntegrationTester $I
-     * @return void
-     */
-    public function deleteMessage(IntegrationTester $I): void
-    {
-        $I->assertTrue(
-            Feeds::delete(Messages::findFirst()->getId(),new Users())
-        );
     }
 }
